requirements.txt
pytesseract
Pillow
pyautogui
pyperclip
tkinter
openai
keyboard